Simple PD working
Test 2 script contains a simple simulink model to control the position of the servo motor using a sine wave as a target and a PD controller to obtain the desired location.

The actual model probably will require the voltage to be the output instead of the angle, so that it is possible to model the servomotor and not just use the built-in functions.

This could be achieved by sending a specific voltage signal to the digital  pin to which the servo motor is connected to.